server/users_micro/requirements.txt
fastapi
#for pandatic
pydantic[email]
uvicorn[standing]
#fo authentication
python-multipart
python-jose[cryptography]
passlib[bcrypt]
#end or auth
#for db
sqlalchemy
psycopg2-binary
psycopg2
#end db
#env file
python-dotenv
requests
pycryptodome